+++ api/src/main/java/com/github/retrooper/packetevents/util/updatechecker/UpdateChecker.java
@@ -19,9 +19,9 @@
 package com.github.retrooper.packetevents.util.updatechecker;
 
 import com.github.retrooper.packetevents.PacketEvents;
-import com.github.retrooper.packetevents.util.AdventureSerializer;
 import com.github.retrooper.packetevents.util.ColorUtil;
 import com.github.retrooper.packetevents.util.PEVersion;
+import com.github.retrooper.packetevents.util.adventure.AdventureSerializer;
 import com.google.gson.JsonObject;
 import net.kyori.adventure.text.format.NamedTextColor;
