+++ api/src/main/java/com/github/retrooper/packetevents/wrapper/status/server/WrapperStatusServerResponse.java
@@ -20,7 +20,7 @@ package com.github.retrooper.packetevents.wrapper.status.server;
 
 import com.github.retrooper.packetevents.event.PacketSendEvent;
 import com.github.retrooper.packetevents.protocol.packettype.PacketType;
-import com.github.retrooper.packetevents.util.AdventureSerializer;
+import com.github.retrooper.packetevents.util.adventure.AdventureSerializer;
 import com.github.retrooper.packetevents.wrapper.PacketWrapper;
 import com.google.gson.JsonObject;
